requirements.txt
python-telegram-bot==20.3
httpx==0.24.1
python-dotenv==1.0.0
sqlite3>=3.35.0